requirements.txt
streamlit
pandas
plotly
lancedb
google-genai
python-dotenv
sentence-transformers
pyarrow
python-dateutil
pypdf
